requirements.txt
flask>=3.0.0
anthropic>=0.30.0
openai>=1.0.0
python-dotenv>=1.0.0
markupsafe>=2.1.0
requests>=2.31.0
